move deadlineNotification file out of models
